projects/views.py
from django.shortcuts import render
from rest_framework import viewsets, permissions
from .models import Project
from .serializers import ProjectsSerializer


class ProjectView(viewsets.ModelViewSet):
    queryset = Project.objects.all()
    serializer_class = ProjectsSerializer
    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
